Fix reactive grid sizing
It only took me 4 hours to find out that the unexpected behavior I was seeing stems from someone somewhere giving grids a minimum height property equal to the combined maximum size of its rows. The same is not true for the width, and it breaks the behavior you'd expect from using the minmax() function in grid-template-rows.