Change margin override logic
It was based on whether the view would overflow with the header added. Since vertical room size responsiveness was fixed in the previous commit, this approach didn't work anymore. The rooms would get smaller upon the header being added to avoid an overflow, prompting the logic to believe that it was safe to override the margin, even though this could still render the header on top of the rooms. To address this, margin overriding safety is now determined by the remaining margin-top of the grid-container *before* the header is added. Thus, the margins will only be overriden if the top spacer margin featured enough room to display the header without affecting the rooms.