Give products a local ID
It is allowed to have multiple of the same type of product in the same room in the same location (e.g. think of a ceiling light fitted out with a couple of the same type of bulb). Thus products in a setup also need a setup-local ID to uniquely identify them