Idea for compatibility filtering
As it turns out my initial approach based entirely off of the "dumb" checkbox-filters was not nearly as flexible as needed for the task at hand. After a few days of brainstorming various approaches I came up with an idea which is similar but should hopefully have enough flexibility, for now.

Compatibility filtering is fully separate from the checkbox-filters. It still uses the native .filter method on arrays, but across various product properties at a time. The "root" of each "filter step" will be a hub and its associated capabilities and restrictions. With a specific hub at hand, we can then filter products based on the control methods which the user uses. If multiple control methods and/or hubs apply, the filtered products will be united, as compatibility to at least one of his interaction paths will be guaranteed at that point.

In regards to the checkbox-filters -- these will not do any of the "heavy lifting" when compatibility filtering is turned out. The compatibility filtering method should disable unwanted checkboxes and try to have the relevant ones ticked for a better UX, but (as stated) these will not and can not represent the actual state of the filtering.

As for devices which can "be made compatible" through e.g. additional purchases -- since these are not compatible per se, they will not appear when compatibility filters are on, at least for now. They would require an entirely different approach to filtering yet again, and I do not have an endless amount of time.